Add alias that uses open to launch Emacs.app
This obviates the need for emacsclient on Mac OS X. Typing 'emacs' calls
'open -a Emacs' to open the file, launching Emacs.app if it isn't
running or just telling it to open the file if it is.

This is, of course, non-portable. I should surround it with a conditional.